python-analysis-v02 executable - expanded console and cmdline args support - refactored code into separate files - added some docstrings
