+++ file_analyzer.py
@@ -1,9 +1,10 @@
 import json
-import string # TODO: delete unnecessary inputs
+import string
 import sys
-from pycodestyle import *
 
-import naming_analysis as naming
+from pycodestyle import BaseReport, Checker, StyleGuide
+
+import naming_analyzer as naming
 
 # TODO: check trailing whitespace? W291, W292, W293, W391
 # TODO: W503, W504? says not enforced by PEP8?
@@ -28,7 +29,14 @@ def error_msg(report, prefix):
     fin = msg[index+len(prefix)+1:]
     return fin
 
-
+# Template function to print information about types of errors found
+# Parameters:
+# - counters: the counters generated by Pycodestyle
+# - report: the file analysis results from Pycodestyle
+# - header: the category of errors being outputed
+# - macro: the file error codes to consider; a sort of "filter" of the
+#   Pycodestyle error codes of interest for this section
+# - clean: template message about no errors of this category found
 def check_errors(counters, report, header, macro, clean):
     print("{} Errors:".format(header))
     has_errs = False
@@ -36,8 +44,8 @@ def check_errors(counters, report, header, macro, clean):
         if err in counters:
             has_errs = True
             print("    {}, {} occurrence{} {}".format(err, counters[err],
-                                                       "s:" if counters[err] > 1 else ": ",
-                                                       error_msg(report, err)))
+                                                      "s:" if counters[err] > 1 else ": ",
+                                                      error_msg(report, err)))
     if not has_errs:
         print("    None. {} statements conform to PEP 8.".format(clean))
 
@@ -46,14 +54,13 @@ def check_errors(counters, report, header, macro, clean):
 # of expected errors passed in through the macro
 # Parameters:
 # - counters: the error count results from Pycodestyle
-# - report: the file analysis results from Pycodestyle
 # - macro: the file error codes to consider; a sort of "filter" of the
 #   Pycodestyle error codes of interest for this section
 # Returns a tuple of dictionary of JSON results and number of errors found
 # Note: This function assumes PEP 8 and Google style guides are the
-# same for the style element being analyzed
-# If the style guides differ, do not use this function
-def json_check_errors(counters, report, macro):
+# same for the style element being analyzed. If the style guides differ,
+# do not use this function
+def json_check_errors(counters, macro):
     temp_dict = {"pep": None, "google": None, "errors": None}
     has_errs = False
     count = 0
@@ -88,31 +95,15 @@ def check_blank_lines(counters, report):
 def check_imports(counters, report):
     check_errors(counters, report, "Import Statement", IMPORT_ERRORS, "Import")
 
-def count_total_file_errors(dict):
-    count = 0
-    for d in dict.keys():
-        if dict[d] is None:
-            pass
-        elif d is NAMES + "_analysis":
-            for type in dict[d].keys():
-                count += dict[d][type]["occurrences"]
-        else:
-            if dict[d]["errors"] is None:
-                pass
-            else:
-                for code in dict[d]["errors"].keys():
-                    count += dict[d]["errors"][code]
-    return count
-
-# Create the dictionary of analysis result values to be converted into JSON output
-def create_json_dict(counters, report, file_name):
+# Create the dictionary of analysis result values to be converted into JSONa
+def create_json_dict(counters, file_name):
     obj = {"total_file_errors": None}
     names = naming.naming_results(file_name)
-    indents = json_check_errors(counters, report, INDENT_ERRORS)
-    tabs = json_check_errors(counters, report, TABS_SPACES_ERRORS)
-    length = json_check_errors(counters, report, LINE_LENGTH_ERRORS)
-    blanks = json_check_errors(counters, report, BLANK_LINE_ERRORS)
-    imports = json_check_errors(counters, report, IMPORT_ERRORS)
+    indents = json_check_errors(counters, INDENT_ERRORS)
+    tabs = json_check_errors(counters, TABS_SPACES_ERRORS)
+    length = json_check_errors(counters, LINE_LENGTH_ERRORS)
+    blanks = json_check_errors(counters, BLANK_LINE_ERRORS)
+    imports = json_check_errors(counters, IMPORT_ERRORS)
     obj[NAMES + "_analysis"] = names[0]
     obj[INDENTS + "_analysis"] = indents[0]
     obj[TABS + "_analysis"] = tabs[0]
@@ -121,27 +112,29 @@ def create_json_dict(counters, report, file_name):
     obj[IMPORTS + "_analysis"] = imports[0]
     obj[ENCODING + "_analysis"] = None
     obj["total_file_errors"] = names[1] + indents[1] + tabs[1] + length[1] + blanks[1] + imports[1]
-    err_counts = {NAMES: names[1], INDENTS: indents[1], TABS: tabs[1], LENGTH: length[1], BLANKS: blanks[1], IMPORTS: imports[1]}
+    err_counts = {NAMES: names[1], INDENTS: indents[1], TABS: tabs[1],
+                  LENGTH: length[1], BLANKS: blanks[1], IMPORTS: imports[1]}
     return obj, err_counts
 
 
 def collect_file_dict_results(file_name):
     # Collect the PEP8 reported errors according to Pycodestyle.
     sg = StyleGuide()
-    # TODO: should be able to "max_line_length=80" in the parens, but not working. add Google line length support
+    # TODO: should be able to "max_line_length=80" in the parens,
+    # but not working. add Google line length support
     breport = BaseReport(options=sg.options)
     # TODO: check file actually exists. otherwise falsely outputting that it is compliant
     quiet_checker = Checker(file_name, report=breport)
     quiet_checker.check_all()
     counters = breport.counters
     # TODO: if a runtime error is thrown (E901, E902), still analyze the rest?
-    js = create_json_dict(counters, breport, file_name)
+    js = create_json_dict(counters, file_name)
     naming.cleanup()
     return js
 
-def check_input(str):
+def check_input(input_str):
     # TODO: double-check this function
-    temp = str
+    temp = input_str
     msg = "That is not a valid file name. Please double-check your input."
     if " " in temp:
         temp = temp.replace(" ", "")
@@ -159,7 +152,6 @@ def main(argv):
     check_input(file_name)
     temp = collect_file_dict_results(file_name)
     res = temp[0]
-    err_counts = temp[1]
     print(json.dumps(res))
     return res
